requirements.txt
psycopg[binary]
paho-mqtt
tenacity